refactor: split admin_console.go (2107 lines) into focused files
Break the monolithic admin_console.go into domain-specific files:
- admin_console.go: struct, constructor, shared types/helpers, constants
- admin_console_dashboard.go: dashboard handler
- admin_console_users.go: user management handlers
- admin_console_orgs.go: organization handlers
- admin_console_clients.go: OAuth client handlers
- admin_console_roles.go: role management handlers
- admin_console_groups.go: group management handlers
- admin_console_sessions.go: session management handlers
- admin_console_events.go: audit events handler
- admin_console_export.go: export/import handlers
- admin_console_oidc.go: OIDC config and social providers

Also replaces 10 literal "Invalid form data." strings with the
existing msgInvalidForm constant (SonarQube S1192 compliance).

Pure refactoring — no behavioral changes.
